chore: refactor rabbit mq consumer and move to seperate files
