+++ zod/slack-schedule.schema.ts
@@ -21,6 +21,7 @@ const slackMsgSchema = z.object({
     messageId: z.string({ message: "property messageId is required" }).optional(),
     text: z.string({ message: "property messageId is required" }),
     channel: z.string({ message: "property messageId is required" }),
+    integrationId: z.string({ message: "property integrationId is required" }),
     message: z.object({
         channelId: z.string({ message: "property channelId is required" }),
         projectId: z.string({ message: "property projectId is required" }),
@@ -49,6 +50,7 @@ const slackMsgScheduleSchema = z.object({
     post_at: z.string({ message: "property post_at is required" }),
     workspaceId: z.string({ message: "property workspaceId is required" }),
     messageId: z.string({ message: "property messageId is required" }).optional(),
+    integrationId: z.string({ message: "property integrationId is required" }),
     message: z.object({
         channelId: z.string({ message: "property channelId is required" }),
         projectId: z.string({ message: "property projectId is required" }),
